+++ src/common/data/data_access.js
@@ -1,4 +1,24 @@
+/// <reference path="../../types/global.d.ts" />
+
+/**
+ * @returns {Fiddler<VotingService.Election>}
+ */
+function getElectionsFiddler() {
+    return Common.Data.Storage.SpreadsheetManager.getFiddler('Elections');
+}
+
+/**
+ * @returns {Fiddler<TokenDataType>}
+ */
+function getTokensFiddler() {
+    return Common.Data.Storage.SpreadsheetManager.getFiddler('Tokens');
+}
+
 Common.Data.Access = {
+    getBootstrapData: () => {
+        const bootStrapFiddler = Common.Data.Storage.SpreadsheetManager.getFiddler('Bootstrap');
+        return bootStrapFiddler.getData();
+    },
     getEmailAddresses: function () {
         const members = Common.Data.Storage.SpreadsheetManager.getFiddler('ActiveMembers').getData();
         const emails = members.map(member => member.Email.toLowerCase());
@@ -11,7 +31,7 @@ Common.Data.Access = {
     getActionSpecs: () => {
         Common.Data.Storage.SpreadsheetManager.convertLinks('Action Specs');
         // We use getDataWithFormulas_ because the Body of an ActionSpec may contain formulas with a URL.
-        const actionSpecsAsArray = Common.Data.Storage.SpreadsheetManager.getDataWithFormulas(Common.Data.Storage.SpreadsheetManager.getFiddler('ActionSpecs'))
+        const actionSpecsAsArray = /** @type {MembershipManagement.ActionSpec[]} */ (Common.Data.Storage.SpreadsheetManager.getDataWithFormulas(Common.Data.Storage.SpreadsheetManager.getFiddler('ActionSpecs')))
         const actionSpecs = Object.fromEntries(actionSpecsAsArray.map(spec => [spec.Type, spec]));
         for (const actionSpec of Object.values(actionSpecs)) {
             let match = actionSpec.Body.match(/=hyperlink\("(https:\/\/docs.google.com\/document\/d\/[^"]+)"/);
@@ -38,5 +58,14 @@ Common.Data.Access = {
             return (member.Email.toLowerCase() === email) ? newMember : member;
         }).dumpValues();
         return true;
+    },
+    isMember:(email) => {
+        email = email.toLowerCase();
+        const members = Common.Data.Storage.SpreadsheetManager.getFiddler('ActiveMembers').getData();
+        return members.some(member => member.Email.toLowerCase() === email);
+    },
+    getElections: () => {
+        const votingData = Common.Data.Storage.SpreadsheetManager.getFiddler('Elections').getData();
+        return votingData;
     }
 }